Compact gate integration tests are rust native
